Removed search bar from nav bar #6
