+++ src/contexts/FirebaseContext.jsx
@@ -1,4 +1,5 @@
 import { getAnalytics } from "firebase/analytics";
+import { getAuth } from "firebase/auth";
 import { getFirestore } from "firebase/firestore";
 import { getStorage } from "firebase/storage";
 import { initializeApp } from "firebase/app";
@@ -12,8 +13,9 @@ export const FirebaseContextProvider = ({ children }) => {
   const app = initializeApp(Config.Firebase());
   const analytics = getAnalytics(app);
   const firestore = getFirestore(app);
+  const auth = getAuth(app);
   const storage = getStorage(app);
-  const value = { app, analytics, firestore, storage };
+  const value = { app, analytics, firestore, storage, auth };
   return (
     <FirebaseContext.Provider value={value}>
       {children}
